Update Index.html
Google Cloud Speech-to-Text API is loaded using the @google-cloud/speech package. The SpeechRecognition object is then created with interimResults set to true so that we can get real-time transcription as the user speaks. The record button starts and stops the recording, and the transcription is displayed in the transcript element.

Note that you will need to set up a Google Cloud account and obtain API credentials in order to use the Speech-to-Text API. You will also need to add the appropriate CORS headers to your backend server to allow the browser to make cross-origin requests to the API.